+++ system/shadowdark.mjs
@@ -73,7 +73,6 @@ Hooks.once("init", () => {
 	CONFIG.Actor.documentClass = documents.ActorSD;
 	CONFIG.ActiveEffect.documentClass = documents.ActiveEffectSD;
 	CONFIG.Item.documentClass = documents.ItemSD;
-	CONFIG.Combat.documentClass = documents.EncounterSD;
 	CONFIG.ChatMessage.documentClass = documents.ChatMessageSD;
 	CONFIG.ActiveEffect.legacyTransferral = false;
